fix #736

+++ turtlebot3_node/src/turtlebot3.cpp
@@ -53,8 +53,8 @@ void TurtleBot3::init_dynamixel_sdk_wrapper(const std::string & usb_port)
 {
   DynamixelSDKWrapper::Device opencr = {usb_port, 200, 1000000, 2.0f};
 
-  this->declare_parameter<uint8_t>("opencr.id");
-  this->declare_parameter<uint32_t>("opencr.baud_rate");
+  this->declare_parameter<int64_t>("opencr.id");
+  this->declare_parameter<int64_t>("opencr.baud_rate");
   this->declare_parameter<float>("opencr.protocol_version");
 
   this->get_parameter_or<uint8_t>("opencr.id", opencr.id, 200);
@@ -142,8 +142,8 @@ void TurtleBot3::add_sensors()
 {
   RCLCPP_INFO(this->get_logger(), "Add Sensors");
 
-  this->declare_parameter<uint8_t>("sensors.bumper_1");
-  this->declare_parameter<uint8_t>("sensors.bumper_2");
+  this->declare_parameter<int64_t>("sensors.bumper_1");
+  this->declare_parameter<int64_t>("sensors.bumper_2");
 
   this->declare_parameter<float>("sensors.illumination");
